% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univeriant.R
\name{Calculate.lew}
\alias{Calculate.lew}
\title{Univarent analysis of Lew data set.}
\usage{
Calculate.lew()
}
\arguments{
\item{na.rm}{A logical scaler. Should missing values be removed? depreciated}
}
\value{
A data frame that returns the mean and standard deviation calculated
    along with the NIST referenced values The data frame is in the same format
    as the data frame result for all NIST Univarent statistic in this package.
    \describe{
       \item{NIST}{The NIST data set utilized. for this function this value
       will be Lew}
       \item{mean}{The R calculated mean from the data set}
       \item{Standard_Deviation}{The R culculated standard deviation from the
       data set}
       \item{NIST_mead}{The NIST value of the mean for the data set, provided
       for reference}
       \item{NIST_Standard_Deviation}{Tne NIST value of of the standard deviation
       for the data set, provided for reference}
       }
}
\description{
\code{lew} returns specifed univarent statistics.
}
\details{
This function calulates specific univariant statistics from the NIST
Lew data set and returns a data frame for further evulation.
}
\examples{
Calculate.lew()

}
