% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univeriant.R
\name{calculate.michelso}
\alias{calculate.michelso}
\title{Univarent analysis of Michelso data set.}
\usage{
calculate.michelso()
}
\value{
A data frame that returns the mean and standard deviation calculated
    along with the NIST referenced values The data frame is in the same format
    as the data frame result for all NIST Univarent statistic in this dataset.
    \describe{
       \item{NIST}{The NIST data set utilized. For this function this value
       will be Michelso}
       \item{mean}{The R calculated mean from the data set}
       \item{Standard_Deviation}{The R culculated standard deviation from the
       data set}
       \item{Autocorrelation_Coefficien}{Calculate population lag-1
       autocorrelation coefficient}
       \item{NIST_mean}{The NIST value of the mean for the data set, provided
       for reference}
       \item{NIST_Standard_Deviation}{Tne NIST value of of the standard deviation
       for the data set, provided for reference}
       \item{NIST_Autocorrelation_Coefficient}{Tne NIST value of of the population
       lag-1 autocorrelation coefficient}
       }
}
\description{
\code{calculate.Michelso} returns specifed univarent statistics.
}
\details{
This function calulates specific univariant statistics from the NIST
Michelso data set and returns a data frame for further evulation.
}
\examples{
calculate.michelso()

}
