% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate.R
\name{calculate.NumAcc4}
\alias{calculate.NumAcc4}
\title{Univariate analysis of NumAcc4 data set.}
\usage{
calculate.NumAcc4()
}
\value{
A data frame that returns the mean, standard deviation and population
    lag-1 autocorrelation coefficient calculated
    along with the NIST referenced values The data frame is in the same format
    as the data frame result for all NIST Univarent statistic in this dataset.
    \itemize{
       \item{\strong{NIST} The NIST data set utilized. For this function this value
       will be NumAcc4}
       \item{\strong{mean} The R calculated mean from the data set}
       \item{\strong{Standard_Deviation} The R culculated standard deviation from the
       data set}
       \item{\strong{Autocorrelation_Coefficien} Calculate population lag-1
       autocorrelation coefficient}
       \item{\strong{NIST_mean} The NIST value of the mean for the data set, provided
       for reference}
       \item{\strong{NIST_Standard_Deviation} The NIST value of of the standard deviation
       for the data set, provided for reference}
       \item{\strong{NIST_Autocorrelation_Coefficient} The NIST value of of the population
       lag-1 autocorrelation coefficient}
       }
}
\description{
\code{calculate.NumAcc4} returns specifed univarent statistics.
}
\details{
This function calulates specific univariate statistics from the NIST
NumAcc4 data set and returns a data frame for further evulation.
}
\examples{
calculate.NumAcc4()
a <- calculate.NumAcc4()

}
