% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking.R
\name{check.univariate}
\alias{check.univariate}
\title{Checks For Correct Results of Univariate Tests.}
\usage{check.univariate()}
\value{
The function returns a vector that shows the version of R the test
was conducted on, the data and time the test was started, the machine
accuracy utilized, the specific data sets run, any noted errors and a count of
errors encountered.
}
\description{
\code{check.univariate} runs each of the univivariate data sets and checks
to that the correct values are returned.
}
\details{
This function runs each of the NIST univariate calculations against the
respective data set. It then compares the results to the given NIST values.
If the valuse do not agree within the expected degree of machine accuracy,
the data set and statistic are logged.

Information on machine accuracy can be found at
\url{http://www.validlab.com/goldberg/paper.pdf} and was calculated with
\code{.Machine$double.ebs ^ 0.5}.
}
\examples{
check.univariate()

a <- check.univariate()

}
