% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariateData.R
\docType{data}
\name{Lottery}
\alias{Lottery}
\title{Lottery NIST Univariate Data Set.}
\format{218 observation of 1 variable:}
\source{
\url{http://www.itl.nist.gov/div898/strd/univ/lottery.html}

#' @examples
Lottery
}
\usage{
Lottery
}
\description{
Dataset of 218 3-digit points
}
\details{
This dataset consists of 218 3-digit numbers (from 000 to 999) resulting from
    the state of Maryland's Pick-3 Lottery. The data was collected for the
    32-week period September 3, 1989 to April 14, 1990. One 3-digit random
    number was drawn per day, 7 days per week for most weeks, but 6 or 5 days
    per week for other weeks. Interesting data-analytic questions involving
    the dataset are 1) are the lottery numbers uniformly distributed? and 2)
    is there serial correlation between lottery numbers?

\tabular{lr}{
    \strong{Level of Difficulty} \tab Lower\cr
    \strong{Variables} \tab 1\cr
    \strong{Observations} \tab 218\cr
    \strong{First observation} \tab 162\cr
    \strong{Expected results} (as certified)\tab \cr
    \strong{Mean} \tab 518.958715596330\cr
    \strong{Standard Deviation} \tab 291.699727470969\cr
    \strong{Population lag-1 autocorrelation coefficient} \tab -0.120948622967393\cr}
}
\keyword{datasets}
