% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univeriant.R
\docType{data}
\name{Mavro}
\alias{Mavro}
\title{Mavro NIST Univarent Data Set.}
\format{50 observation of 1 variable:}
\source{
\url{http://www.itl.nist.gov/div898/strd/univ/Mavro.html}
}
\usage{
Mavro
}
\description{
Dataset of 50 points
}
\details{
This "real world" dataset is the result of a study by Radu Mavrodineaunu, a
    chemist at the National Institute of Standards & Technology (NIST). The
    purpose of the study was to determine a certified transmittance value
    that may be attached to the particular of filter under study. The 50
    transmittance valuess were collected equi-spaced in time at a sampling
    rate of 10 observations per second.

\tabular{lr}{
    \strong{Level of Difficulty} \tab Lower\cr
    \strong{Variables} \tab 1\cr
    \strong{Observations} \tab 50\cr
    \strong{First observation} \tab 2.00180\cr
    \strong{Expected results} (as certified)\tab \cr
    \strong{Mean} \tab 2.00185600000000\cr
    \strong{Standard Deviation} \tab 0.000429123454003053\cr
    \strong{Population lag-1 autocorrelation coefficient} \tab 0.937989183438248\cr}
}
\keyword{datasets}
