% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariateData.R
\docType{data}
\name{Michelso}
\alias{Michelso}
\title{Michelso NIST Univariate Data Set.}
\format{100 observation of 1 variable:}
\source{
\url{http://www.itl.nist.gov/div898/strd/univ/Michelso.html}
}
\usage{
Michelso
}
\description{
Dataset of 100 points
}
\details{
This "real world" dataset is the result of the classic study conducted by
    Michelson on the speed of light in air in 1879. The response variable is
    speed of light (in millions of meters per second). The data was included
    as part of a larger study by Dorsey, Ernest N. (1944) on the velocity of
    light as reported in the Transactions of the American Philosophical Society.

\tabular{lr}{
    \strong{Level of Difficulty} \tab Lower\cr
    \strong{Variables} \tab 1\cr
    \strong{Observations} \tab 100\cr
    \strong{First observation} \tab 299.85\cr
    \strong{Expected results} (as certified)\tab \cr
    \strong{Mean} \tab 299.852400000000\cr
    \strong{Standard Deviation} \tab 0.0790105478190518\cr
    \strong{Population lag-1 autocorrelation coefficient} \tab 0.535199668621283\cr}
}
\keyword{datasets}
