% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univeriant.R
\docType{data}
\name{NumAcc1}
\alias{NumAcc1}
\title{NumAcc1 NIST Univarent Data Set.}
\format{3 observation of 1 variable:}
\source{
\url{http://www.itl.nist.gov/div898/strd/univ/NumAcc1.html}
}
\usage{
NumAcc1
}
\description{
Dataset of 3 8-digit points
}
\details{
This generated/fabricated dataset consists of three 8-digit integers
    differing only in the least significant digit. The data set is: 10000002,
    10000001, and 10000003. By construction, this data set has sample
    mean = 10000002 (exact); sample standard deviation = 1 (exact); and
    sample autocorrelation coef. = -0.5 (exact). The construction was carried
    out based on considerations described by Simon, Stephen D. and Lesage,
    James P. (1989): Assessing the Accuracy of ANOVA Caluclations in Statistical
    Software", Computational Statistics \& data Analysis, 8, pp. 325-332

\tabular{lr}{
    \strong{Level of Difficulty} \tab Average\cr
    \strong{Variables} \tab 1\cr
    \strong{Observations} \tab 1001\cr
    \strong{First observation} \tab 10000001\cr
    \strong{Expected results} (as certified)\tab \cr
    \strong{Mean} \tab 10000002 (exact)\cr
    \strong{Standard Deviation} \tab 1 (exact)\cr
    \strong{Population lag-1 autocorrelation coefficient} \tab -0.5(exact)\cr}
}
\keyword{datasets}
