% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariateData.R
\docType{data}
\name{NumAcc2}
\alias{NumAcc2}
\title{NumAcc2 NIST Univariate Data Set.}
\format{101 observation of 1 variable:}
\source{
\url{http://www.itl.nist.gov/div898/strd/univ/numacc2.html}

#' @examples
NumAcc2
}
\usage{
NumAcc2
}
\description{
Dataset of 1001 2-digit points
}
\details{
This generated/fabricated dataset consists of 1001 2-digit floating-point
     values: a single 1.2, followed by 500 pairings of 1.1 and 1.3. By
     construction, this data set has sample mean = 1.2 (exact); sample
     standard deviation = 0.1 (exact); and sample autocorrelation coef. = -0.999
     (exact). The construction was carried out based on considerations
     described by Simon, Stephen D. and Lesage, James P. (1989): Assessing
     the Accuracy of ANOVA Caluclations in Statistical Software", Computational
     Statistics \& data Analysis, 8, pp. 325-332.

\tabular{lr}{
    \strong{Level of Difficulty} \tab Average\cr
    \strong{Variables} \tab 1\cr
    \strong{Observations} \tab 1001\cr
    \strong{First observation} \tab 1.2\cr
    \strong{Expected results} (as certified)\tab \cr
    \strong{Mean} \tab 1.2 (exact)\cr
    \strong{Standard Deviation} \tab 0.1 (exact)\cr
    \strong{Population lag-1 autocorrelation coefficient} \tab -0.999 (exact)\cr}
}
\keyword{datasets}
