% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariateData.R
\docType{data}
\name{NumAcc4}
\alias{NumAcc4}
\title{NumAcc4 NIST Univariate Data Set.}
\format{1001 observation of 1 variable:}
\source{
\url{http://www.itl.nist.gov/div898/strd/univ/numacc4.html}
}
\usage{
NumAcc4
}
\description{
Dataset of 1001 9-digit points
}
\details{
This generated/fabricated dataset consists of 1001 9-digit floating-point
    values: a single 10000000.2, followed by 500 pairings of 10000000.1 and
    10000000.3. By construction, this data set has sample mean = 10000000.2
    (exact); sample standard deviation = .1 (exact); and sample autocorrelation
    coef. = -0.999 (exact). The construction was carried out based on
    considerations described by Simon, Stephen D. and Lesage, James P. (1989):
    Assessing the Accuracy of ANOVA Caluclations in Statistical Software",
    Computational Statistics \& data Analysis, 8, pp. 325-332.

\tabular{lr}{
    \strong{Level of Difficulty} \tab Higher\cr
    \strong{Variables} \tab 1\cr
    \strong{Observations} \tab 1001\cr
    \strong{First observation} \tab 100000000.2\cr
    \strong{Expected results} (as certified)\tab \cr
    \strong{Mean} \tab 10000000.2 (exact)\cr
    \strong{Standard Deviation} \tab 0.1 (exact)\cr
    \strong{Population lag-1 autocorrelation coefficient} \tab -0.999 (exact)\cr}
}
\examples{
NumAcc4
}
\keyword{datasets}
