% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univeriant.R
\docType{data}
\name{PiDigits}
\alias{PiDigits}
\title{Pi Digits NIST Univarent Data Set.}
\format{218 observation of 1 variable:}
\source{
\url{http://www.itl.nist.gov/div898/strd/univ/pidigits.html}
}
\usage{
PiDigits
}
\description{
Dataset of 5000 points
}
\details{
This dataset consists of the first 5000 digits of the mathemtatical constant
    pi (= 3.1415926535897932384...). These 5000 digits were reported in
    Mathematics of Computation, January 1962, page 76. Interesting
    number-theoretic questions involving pi digits are 1) are the digits
    uniformly distributed? and 2) is there serial correlation between
    successive digits?

\tabular{lr}{
    \strong{Level of Difficulty} \tab Lower\cr
    \strong{Variables} \tab 1\cr
    \strong{Observations} \tab 5000\cr
    \strong{First observation} \tab 3\cr
    \strong{Expected results} (as certified)\tab \cr
    \strong{Mean} \tab 4.53480000000000\cr
    \strong{Standard Deviation} \tab 2.86733906028871\cr
    \strong{Population lag-1 autocorrelation coefficient} \tab -0.00355099287237972\cr}
}
\keyword{datasets}
