% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RQualification.R
\docType{package}
\name{RQualification}
\alias{RQualification}
\alias{RQualification-package}
\title{RQualification: A package to compute NIST Statistical Data Sets.}
\description{
The RQualification package provides three groups of functions and data to
verify that R is preforming as expected against data sets from the national
Institute of Standards and Techinology (NIST).
}
\section{Checking Functions}{

   \itemize{
      \item{\link{check.univariate} Checks For Correct Results of Univariate
      Tests.}
      }
}

\section{Univariate Statistics}{

   \itemize{
      \item{\link{calculate.lew} This function calulates specific Univariate
      statistics the NIST \link{Lew} data set and returns a data frame for
      further evulation.}
      \item{\link{calculate.lottery} This function calulates specific Univariate
      statistics the NIST \link{Lottery} data set and returns a data frame for
      further evulation.}
      \item{\link{calculate.mavro} This function calulates specific Univariate
      statistics the NIST \link{Michelso} data set and returns a data frame for
      further evulation.}
      \item{\link{calculate.mavro} This function calulates specific Univariate
      statistics the NIST \link{Mavro} data set and returns a data frame for
      further evulation.}
      \item{\link{calculate.NumAcc4} This function calulates specific Univariate
      statistics the NIST \link{NumAcc4} data set and returns a data frame for
      further evulation.}
      \item{\link{calculate.NumAcc3} This function calulates specific Univariate
      statistics the NIST \link{NumAcc3} data set and returns a data frame for
      further evulation.}
      \item{\link{calculate.NumAcc2} This function calulates specific Univariate
      statistics the NIST \link{NumAcc2} data set and returns a data frame for
      further evulation.}
      \item{\link{calculate.NumAcc1} This function calulates specific Univariate
      statistics the NIST \link{NumAcc1} data set and returns a data frame for
      further evulation.}
      \item{\link{calculate.PiDigits} This function calulates specific Univariate
      statistics the NIST \link{PiDigits} data set and returns a data frame for
      further evulation.}
}
}

\section{Linear Regression}{

under development
}

\section{Non-Linear Regression}{

under development
}

