% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univeriant.R
\docType{data}
\name{Lew}
\alias{Lew}
\title{Lew NIST Univarent Data Set}
\format{200 observation of 1 variable:}
\source{
\url{http://www.itl.nist.gov/div898/strd/univ/lew.html}
}
\usage{
Lew
}
\description{
INfo on data set
}
\keyword{datasets}
